+++ runtime/src/pallets/mod.rs
@@ -27,3 +27,6 @@ pub use xcmp_queue::*;
 
 pub mod dmp_queue;
 pub use dmp_queue::*;
+
+pub mod session;
+pub use session::*;
